@@ -1,6 +1,9 @@
 Rails.application.config.after_initialize do
   Bullet.enable = true
-  Bullet.alert = true
-  Bullet.rails_logger = true
+  Bullet.bullet_logger = true
   Bullet.console = true
+  Bullet.rails_logger = true
+  Bullet.add_footer = true
 end if defined?(Bullet)
+
+
